requirements.txt
django==4.1.1
djangorestframework